+++ lib/items/general/listing.dart
@@ -51,7 +51,7 @@ class _ListingState extends State<ListingBuilder> {
     _scrollController.addListener(_onScroll);
     if (widget.refresh) {
       listingBloc.load();
-    } else if (listingBloc.currentSubmissions.isEmpty) {
+    } else if (listingBloc.loadedData.isEmpty) {
       listingBloc.load();
     }
     listingBloc.jumpToTopStream
@@ -84,79 +84,59 @@ class _ListingState extends State<ListingBuilder> {
     }
     slivers.add(SliverList(
         delegate: SliverChildBuilderDelegate((context, index) {
-      var item = listingBloc.currentSubmissions.keys.toList()[index];
-      return widget.builder(context,
-          ListingSnapShot(listingBloc.currentSubmissions[item], index));
-    }, childCount: listingBloc.currentSubmissions.length)));
+      var item = listingBloc.loadedData.keys.toList()[index];
+      return widget.builder(
+          context, ListingSnapShot(listingBloc.loadedData[item], index));
+    }, childCount: listingBloc.loadedData.length)));
     return slivers;
+  }
+
+  Widget buildScaffold(){
+    return Scaffold(
+      body: LoadingScreen(),
+    );
   }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      resizeToAvoidBottomPadding: false,
-      body: new Stack(
-        children: <Widget>[
-          new StreamBuilder(
-            stream: listingBloc.resultsStream,
-            initialData: false,
-            builder: (BuildContext context, AsyncSnapshot<bool> snapshot) {
-              if (snapshot.hasError) {
-                if (widget.error != null) {
-                  return widget.error(snapshot.error);
-                } else {
-                  return RetryWidget(
-                    message: Text(snapshot.error.toString()),
-                    onTap: () => listingBloc.load(refresh: true),
-                  );
-                }
-              }
-              if (listingBloc.currentSubmissions.isNotEmpty) {
-                return RefreshIndicator(
-                  onRefresh: () => widget.listingBloc.load(refresh: true),
-                  child: Scrollbar(
-                    child: CustomScrollView(
-                      controller: _scrollController,
-                      slivers: buildSlivers(),
-                    ),
-                  ),
-                );
-              } else if (snapshot.data == false) {
-                return widget.loading ?? LoadingScreen();
-              } else if (snapshot.data == null) {
-                return widget.empty ??
-                    RetryWidget(
-                        onTap: () => widget.listingBloc.load(
-                              refresh: true,
-                            ));
-              } else {
-                return Container(
-                  width: 0.0,
-                  height: 0.0,
-                );
-              }
-            },
-          ),
-          Padding(
-            padding: const EdgeInsets.only(top: 8.0),
-            child: new StreamBuilder(
-              stream: listingBloc.isLoadingStream,
-              initialData: false,
-              builder: (BuildContext context, AsyncSnapshot<bool> snapshot) {
-                if (snapshot.data &&
-                    listingBloc.currentSubmissions.isNotEmpty) {
-                  return LinearProgressIndicator();
-                } else {
-                  return Container(
-                    width: 0.0,
-                    height: 0.0,
-                  );
-                }
-              },
+    return new StreamBuilder(
+      stream: listingBloc.resultsStream,
+      initialData: false,
+      builder: (BuildContext context, AsyncSnapshot<bool> snapshot) {
+        if (snapshot.hasError) {
+          if (widget.error != null) {
+            return widget.error(snapshot.error);
+          } else {
+            return RetryWidget(
+              message: Text(snapshot.error.toString()),
+              onTap: () => listingBloc.load(refresh: true),
+            );
+          }
+        }
+        if (listingBloc.loadedData.isNotEmpty) {
+          return RefreshIndicator(
+            onRefresh: () => widget.listingBloc.load(refresh: true),
+            child: Scrollbar(
+              child: CustomScrollView(
+                controller: _scrollController,
+                slivers: buildSlivers(),
+              ),
             ),
-          )
-        ],
-      ),
+          );
+        } else if (snapshot.data == false) {
+          return widget.loading ?? LoadingScreen();
+        } else if (snapshot.data == null) {
+          return widget.empty ??
+              RetryWidget(
+                  onTap: () => widget.listingBloc.load(
+                    refresh: true,
+                  ));
+        } else {
+          return CustomScrollView(
+            slivers: <Widget>[widget.sliverAppBar],
+          );
+        }
+      },
     );
   }
 }