+++ lib/utils/dialogs.dart
@@ -245,4 +245,19 @@ class Dialogs {
               ],
             ));
   }
+  static Future<bool> showLogoutConfirmationDialog(BuildContext context) async {
+    return await showDialog(context: context,
+    builder: (context) => AlertDialog(
+      title: Text("Confirm Sign out"),
+      content: Text("Are you sure?"),
+      actions: <Widget>[
+        FlatButton(
+            onPressed: () => Navigator.pop(context, false),
+            child: Text("No")),
+        FlatButton(
+            onPressed: () => Navigator.pop(context, true),
+            child: Text("Yes"))
+      ],
+    ));
+  }
 }